pyproject.toml
[project]
name = "chess-teacher"
version = "0.1.0"
description = "A chess teaching application"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-chess>=1.999",
    "chromadb>=0.6",
    "pydantic-settings>=2",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that require external services (deselect with '-m \"not integration\"')",
]
